lab4/api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
kafka-python==2.0.2
pydantic==2.6.3
